renamed StepDescription
